Starting to look like a file system
Now as you type, if you type a valid file path, you get an updated list of files in the second box.
Of course you can't do anything with the files yet; I have a long way to go for that.  Here's what
I'm thinking of as of right now:
    - Arrow keys to navigate the list
    - Delete key to delete a file (obviously) (maybe down the road add a confirm dialog, lol... for
        now this is hardcore Linux - it's like running rm or rmdir :D).
    - P to get the file properties (and maybe edit them down the road)
    - Enter to open the file (in nano for now)
    - O for an "open with" menu, waaaay down the road.
    - CTRL+Z and CTRL+Y to undo/redo file delete
    - And F1 for a menu with all this info, maybe even a menu bar

If I can accomplish even half of these things, I'm well on my way to
a pretty complete file manager.  Of course it's gonna be my biggest
C project ever, probably involving linked lists and dynamic memory
management and all that stuff I understand, but rarely use (lol).
And even if this doesn't work out how I want, it'll give me some
serious practice with important skills like debugging with GDB, in
the future maybe figure out make and get it to use that, etc. :D

And speaking of debugging, I've hit my first bug!  Yay?  Seriously
though, this one's a doozy: for some reason, eSpeak is not speaking.
Since accessibility is such a key priority for this project, this
is an important one.