Added a license (The Un-License)
I am a code monkey, not a legal beagle.  I don't claim to know what I'm doing
in this area, but I want this project to be as freed as possible.  I say the
word "freed" to mean libre (free as in freedom) and distinguish it from
"free" as in gratis (as in free pizza).  IMO we shouldn't need to clarify
which type of free we're talking about by using another language (lol).  I
am fluent in Spanish so I understand libre vs. gratis, but is our language
really that limited?  lol I just think "freed" makes more sense in this
context.  Anyway, i want this project to be as freed as possible, so I'm not
sure what ncurses and espeak are (except that they're open-source).  I know
ncurses has been around for decades and its copyright history is ridiculously
over-complicated, and I think eSpeak might be GPL or LGPL.  Either way, I'm
not even sure if the two licenses are compatible, but I'm using both anyway,
so the Unlicense just makes the most sense IMO.  If you think this project
needs to be licensed under something else, fell free to try and explain it
to me in English.  I'm open to changing it if it has to be changed.