Started on #3
The focus indicator is working, but I realize there is no apparent way
(to a visual user) to move the focus back to the text box; I think this
is easily solved by an instruction in the help file, but it might be a
good idea to consider putting the list in the up/down arrow sequence.
Of course, I'd have to make sure the speech gives this info to users,
and maybe I should set it up so users can Tab between the text box and
the list box.