Gettin' there...
I got it to not overflow, but now I've got a math problem on my hands.
I'm trying to figure out how to get this thing to scroll.  For example,
if you arrow down the root folter ("/") you can arrow down past the
last visible item on the list.  eSpeak speaks the things that come after
it, but it's not visible (and not all users will use the speech).  So
at this point I'm thinking maybe a paging system of some sort, but even
then the implementation details are gonna be tricky.  Another thought
that occurred to me is to add a "previous" pointer to my linked list,
which is equally tricky but also doable with enough work.  So I think
at this point I'm gonna call it a night and step away from the code.
Stupidly-complex code problems are easier to solve when my mind is off
duty, I think. :D