Started on file delete dialog
This took a pretty major move, with a different while loop for each window.
The actual file-deleting functionality is not there yet, but at least it
shows the dialog.  I would have preferred to use Enter and Escape, and I
might still try to go that route.  But for now, Y or N works.